refactor(model-providers): consolidate services into a single directory
Three sibling locations existed for one cohesive concern — credentials
CRUD, OAuth flow, pairings, refresh worker, token resolver, registry —
spread across `services/model-provider-credentials.ts` (file),
`services/model-providers/` (registry only), and
`services/oauth-model-providers/` (everything else). One of the doc
comments in `credentials.ts` even pointed at the wrong sibling, a
classic side-effect of the layout suggesting separate domains for
what is one domain.

Collapses to one `services/model-providers/` directory:
  credentials.ts        (← services/model-provider-credentials.ts)
  oauth-flow.ts         (← services/oauth-model-providers/oauth-flow.ts)
  pairings.ts           (← services/oauth-model-providers/pairings.ts)
  refresh-worker.ts     (← services/oauth-model-providers/refresh-worker.ts)
  token-resolver.ts     (← services/oauth-model-providers/token-resolver.ts)
  registry.ts           (unchanged)

Integration test files rename to match: `oauth-model-providers-*.test.ts`
→ `model-providers-*.test.ts` (the `oauth-` prefix only made sense when
there was a dedicated `oauth-model-providers/` dir to disambiguate
from).

URL paths (`/api/model-provider-credentials`, `/api/model-providers-oauth`),
permission scopes (`model-provider-credentials:*`), frontend query keys,
and `routes/model-provider-credentials.ts` / OpenAPI path files
deliberately do NOT move — those are external contracts.

Pure layout change: 20 files touched, no logic edits. `bun run check`
green; 115/115 touched integration tests + 14/14 codex unit tests pass.
