+++ apps/api/src/lib/boot.ts
@@ -48,7 +48,7 @@ import { listOrphanRunIds } from "../services/state/runs.ts";
 import { synthesiseFinalize } from "../services/run-event-ingestion.ts";
 import { initScheduleWorker } from "../services/scheduler.ts";
 import { initInlineCompactionWorker } from "../services/inline-compaction.ts";
-import { initOAuthModelRefreshWorker } from "../services/oauth-model-providers/refresh-worker.ts";
+import { initOAuthModelRefreshWorker } from "../services/model-providers/refresh-worker.ts";
 import { initCancelSubscriber } from "../services/run-tracker.ts";
 import { startRunWatchdog } from "../services/run-watchdog.ts";
 import { getOrchestrator } from "../services/orchestrator/index.ts";
