+++ apps/api/src/services/model-providers/pairings.ts
