+++ packages/db/src/schema/organizations.ts
@@ -135,7 +135,7 @@ export const orgProxies = pgTable(
  * table was dropped. `provider_id` is a free-text registry key (e.g. "codex",
  * "openai") — NOT a FK to `packages.id`. Inference wire format and default
  * base URL are read from the platform registry
- * (`apps/api/src/services/oauth-model-providers/registry.ts`) keyed by
+ * (`apps/api/src/services/model-providers/registry.ts`) keyed by
  * `provider_id`. `base_url_override` is honored only for providers whose
  * registry entry has `baseUrlOverridable: true` (e.g. "openai-compatible").
  *
